cache: add noDiskWrite option for cache.set()
This fixes a race condition (and just excessive CPU/disk activity)
caused by the `cache.set()` operation that is at the end of
`cache.fetch()`. Since we had just read that data out of the file on
disk, writing it back is excessive. In rare cases, it can also lead to
cache corruption, if a file is clobbered while being read.