registry-client: s/306/304/
A typo led to treating 306 as the 'unchanged' status code, which of course
it is not.

@ruyadorno noticed that it was making full fetches for packuments that
ought to have been in the cache, and indeed, this was the reason.

This also improves the benchmarks somewhat, since it is now actually
benefitting from the fact that a cached entry can be re-used if the etag
has not changed.