graph: fix reading actual files
Fixes `src/build-actual.ts` to read from a file structure that should
be closer to the final product.

Also adds small tweaks to the `Package` and its inventory, allowing for
packages metadata to be updated, such that the resulting graph from
reading a lockfile and the actual tree can share the same inventory that
gets hydrated as more metadata becomes available.