graph: add lockfile module
Adds a lockfile module that can save and load lockfiles representing
install graphs.

Lockfile graphs are special in the sense that they're a much more
minimalistic representation of the install, having no manifest
data it relies on info that is hydrated from the lockfile, currently
only a `DepID`, `Integrity` and `resolved` values for packages.

This changeset also includes a few refactors that complements the
integration of the new concepts brought forward by the lockfile:

- Added `Node.integrity` and `Node.resolved` values
- `Node.manifests` are now an optional property
- `Node.location` now defaults to relative paths
- `Node.location` basename now fallback to `hydrate(id).name`
- Added `Node.setResolvedFromId` method to infer `Node.resolved` values
  from hydrated id info
- Removed unused `src/graph/src/realpath.js` file
- Added resolved and integrity values to human readable output viz
- Added a `Graph.edges: Set<Edge>` property to keep track of edges
- Added `Graph.#config` to keep track of @vltpkg/config values
- Added `Node.#config` so that @vltpkg/dep-id.hydrate can use it

PR-URL: https://github.com/vltpkg/vltpkg/pull/52
Credit: @ruyadorno
Close: #52
Reviewed-by: @isaacs