infer Manifest and Packument types, check validity
Running the checks in `asPackument` and `asManifest` does add some CPU
cost, which we may want to do away with at some point. But unless it
becomes a problem, having the additional data validation will prevent a
lot of crashes down the line.

Ideally, we can do this data checking _only_ when fetching the
packument/manifest in `@vltpkg/package-info`, and from there on, just
trust the types, so the cost will be minimal and the benefit will spread
throughout all the other points where we handle these objects.