cache: simplify dir structure, no unixy nesting
The only reason to have done this previously was the lack of a way to
stream over the entries in a directory. But node has had 'opendir' for a
long time now, and we may as well use it.

Note: the cache directory is a bit of a footgun with this change if you
ever _do_ try to 'readdir()' it. So don't do that. Let Cache handle the
cache directory for you, that's what it's for.

This simplification will make it easier to add integrity-based content
addressable caching as well.