cache: fetch/set using a provided Integrity string
The Integrity value can be passed via `fetchOptions.context`, or in the
`set()` options.

If the integrity provided is not a valid Integrity string, then it is
ignored.

Integrity values are not calculated or verified. The caller must do this
check, if desired.

Note that the integrity provided to `cache.fetch()` or `cache.set()`
does _not_ typically match the calculated integrity of the `CacheEntry`
object being cached. Typically, the integrity is related to the body of
the response that the `CacheEntry` represents.

Fix: #29