chore: unify devDependencies into workspace root
Our various devDeps were getting very out of sync with one another,
resulting in many unnecessary copies of @types/node, tshy, and a few
others.

This moves all external devDeps to the top-level workspace root, so that
they can be managed in a single location. *Internal* devDependencies
should still be listed in the individual packages, so that they can be
built in the correct order.

While moving things around, also noticed that `boolean` is not part of
`JSONField`, and it should be, and tightened the key types of
`ExportsSubpaths` objects because why not.