spec: complete the set of allowed workspace specifiers
name@workspace: <-- same as name@workspace:*
x@workspace:y@* <-- alias workspace 'y' as dep named 'x'
x@workspace:y@  <-- same as x@workspace:y@*

The @ sign is needed if a workspace name is specified, otherwise we can't tell
that it's a workspace identifier. It would be possible to omit if we dropped
support for a semver range in that position, perhaps worth considering.