dep-id: hydrate registry specs properly
If a name is provided to the hydration method, use that. If not, fall
back to the (intentionally invalid) name `'(unknown)'`.

Make the DepID for a workspace _only_ capture the workspace name/path,
since the workspaceSpec is only relevant when it comes time to swap out
for publication. Where it concerns dependencies (and thus lockfiles and
the contents of our dep graph), "depend on workspace X" has the same
effect regardless of the spec. It's always just a symlink to the
workspace directory.

With the change to registry specs, the following hydrations can be
expected:

- With `name` = `'x'`

  - `registry;;x@*` -> `x@*`
  - `registry;;y@*` -> `x@npm:y@*`
  - `registry;foo;x@*` -> `x@foo:x@*`
  - `registry;foo;y@*` -> `x@foo:y@*`

- With name undefined:

  - `registry;;x@*` -> `x@*` (use name from spec)
  - `registry;;y@*` -> `x@npm:y@*` (use name from spec)
  - `registry;foo;x@*` -> `(unknown)@foo:x@*` (Must provide _some_ name,
    because no name on initial spec, only final subspect)
  - `registry;foo;y@*` -> `(unknown)@foo:y@*`