config: features and fixes to support CLI
Cleans up the API surface and provides more affordances for getting,
setting, and deleting config entries.

This also defines the set of top-level vlt commands, which feels like a
bit of an overly entangled relationship with the CLI module itself.

A subsequent commit, once the cli module is in place, will pull the
definitions and command names out of this module, and let them live in
the CLI where they belong.