registry-client: v8.serialize cached JSON objects
Most of the time that we fetch JSON from the registry, we immediately
JSON.parse() the results. Writing JSON text back to the cache is
inefficient, as this is slower to parse and slower to generate.

The serializedHeader and serialize/deserialize methods are isolated in
their own module, so that we can swap them out the appropriate
mechanisms based on the JavaScript engine.

This necessitated the addition of a `CacheEntry.setHeader()` method,
which is now also used to set the `content-length` and
`content-encoding` to keep them in sync with body encoding/length
changes. So those headers are no longer misleading.

Lastly, this fixes a purely cosmetic but very confusing bug where it was
failing to count by 2's properly in `CacheEntry.#headersAsObject`.

A prior attempt at this idea stored the serialized buffer as the
`body`, with a `content-type` header indicating that it's a
serialization rather than JSON text. However, this is not good, since
the serialization can change between JS engine major versions, and thus
can make the cache invalidated (or appear corrupted, when it fails to
deserialize).

To address that, this implementation stores the serialized data to a
custom header instead. This way, the serialized version of the response
body can be cached specific to each platform major version, since
serialization can change between major versions. If we don't have a
serialized copy for the current platform version, we just parse the text
as we normally would, but still benefit from having the cache available
to skip the HTTP request.

When switching between major versions, for example if the vlt cache
persists between CI runs, then multiple serialized copies may be stored
on the entry. This bloats the single entry, but avoids bloating the
cache as a whole by making the platform version a part of the cache key,
or invalidating the cache entry if the cached serialized object cannot
be safely deserialized.

Fix: #73