graph: add Diff class
Given two graphs, a `Diff` object will tell you how to turn one into the
other, by listing the nodes and edges that need to be added or removed.

Typically, the `graphFrom` will be the result of a `actual.load()`, but
the `Diff` class is agnostic as to what the graphs actually represent.