graph: use @vltpkg/satisfies, other improvements
Making the build ideal process able to re-use existing nodes in the
graph, as well as the changes to remove `registry` from the start of
most DepID strings, turned out to require quite a few other changes as
well.

For convenience more than anything, the `lockfileData` is now the
standard `util.inspect()` result for a Graph object, which is very nice
for debugging in the node REPL.

In order to have a fast lookup for matching nodes, a `graph.nodesByName`
map is maintained. From there, we can quickly find the nodes with a
given name, and outside of extreme edge cases, this is going to be a
small enough set that it's reasonable to check them one by one for a
match. In order to prevent having to look it up multiple times, any time
such a resolution is found, the reference is cached, so that it becomes
an O(1) operation after the first hit for any given spec.

`devDependencies` are now added for importers, `git` deps, and nothing
else. All other dependency types *are* included (ie, peerDeps are added
to the ideal graph).

Failure to load the manifest for an optional or peerOptional dep is
ignored, and the system will proceed in that case without raise an
error.

Because it's needed so often, and has to be used by the Graph anyway,
the `monorepo` and `projectRoot` fields are exposed on the object.

The `Node.name` property is replaced with a getter, so that rather than
`node.name || node.id` everywhere, we can just use `node.name` and it'll
fallback appropriately.

Rather than separate `removeNode` and `removeEdge` methods, just a
single `removeNode` method is now there, which removes the node, removes
all of its `edgesOut`, and marks any `edgesIn` as missing.

In order to not create misleading graph representations when importers
link to nodes that are not where vlt would place them (for example,
repos using pnpm, such as this one), the `lockfileData` now tracks the
`node.location` for any nodes, if they are not in the "normal" vlt
folder location. There's an open question as to whether we should keep
those when we reify, but it seems most sensible for `vlt install` to
relocate any nodes in `node_modules` that are not in their "correct"
place. A subsequent commit will have `buildIdeal` move them into their
correct place, so that `graph.Diff` can note them as changed.

Lastly, moved the Graph classes to name their options object `options`
instead of `config`, following the naming convention we decided upon
where "config" means "a thing a user sets in a file, cli, or env", and
"options" is "a data object passed to a function or class". (This is
obviously arbitrary, as these words are synonyms, but it helps keep
things a bit more organized.)