graph: refactor lockfile edge storage, store node flags
This makes two changes to the lockfile data structure.

The first, is that the 'dev' and 'optional' flags for Node objects are
stored in a numeric entry set to either 0, 1, 2, or 3, which can then be
bit-wise set on the Node objects when constructed, which is about the
most dense and performant way there is to store a small number of
booleans.

Second, edges are now stored such that it is just simply impossible to
ever have multiple edges that attempt to write symbolic links to the
same target path on disk (an error corrected in a recent commit). While
it is of course best to not have such duplicate edges in the first
place, this ensures that the data structure itself prevents this.

A corrollary of this is that it's no longer necessary to mutate the
edges list to put each entry on its own line, as JSON.stringify will do
that automatically. It is ever so slightly slower to parse strings
rather than just have them stored as a JSON tuple, but the simplicity
and correctness defense in depth is more than valuable enough to justify
the minimal one-time expense.

The flags on the Node object will be used in a subsequent update to
track which nodes are dev and/or optional during `ideal.build()` and
`actual.load()`, which is required to properly handle optional failures,
include/omit/query different dependency types, and so on. (Note: simply
looking at the edge types coming into a node is not enough; a prod
dependency of an optional dependency is optional, for example.)

Fix: https://github.com/vltpkg/vltpkg/issues/107