enforce catalog prefix for shared prod deps
As we develop commands and workspaces, it will become important to make
sure we default to always using the same major version of shared
dependencies unless we explicitly opt-out of that behavior.

This updates `scripts/consistent-package-json.js` to look through all
prod deps like it was already doing for dev deps. If the script
encounters two conflicting versions of the same dep used across
workspaces it will error asking us to either update the dep to be
compatible across all workspaces OR to manually add an entry to the
script to skip cataloging that dependency.

This script found one incompatible use of `polite-json@4` in
`package-json` which has now been updated to v5.

Also the script had a bug where devDeps would always be set to catalog:
instead of workspace: even if they were internal workspaces. This fixes
that.