graph: handle optional dep failures in ideal build
This adds the ability of the ideal graph build process to fully handle
ideal node resolution failures wherever they occur within the optional
subgraph.

That is, if a non-optional dependency (direct or transitive) of an
optional dependency fails, then the affected nodes will be removed, but
otherwise the ideal graph construction will continue. If any nodes are
so removed, this *may* leave some portion of the graph disconnected, so
a garbage collection mark-and-sweep is performed at the end of the
process if this happened.

In the process of doing this, a bug was uncovered, where a node could
end up not being properly un-marked as optional. Consider:

```
root
  -OPT-> a -> b -> c -> d
  -> x -> b
```

If root resolves `a` first, it'll be marked as optional, and this flag
will transmit to all of the deps added to support it (`b`, `c`, and
`d`).

Then some time later, `x` is resolved, and has a direct dependency on
`b`. We were properly flagging that `b` is no longer optional. However,
this was *not* being transmitted down the `edgesOut` to all
optional-flagged dependencies of the now no-longer-optional dependency
node, leaving `c` and `d` in an `optional` state incorrectly.

To address this, the `dev` and `optional` flags are now getter/setter
properties, which transmit to all flagged downstream dependencies when
being set `false`.