graph: always gc() after building ideal
This is needed any time we add or change pretty much anything, since we could
always end up making a change that strands some dependencies, even if an optional
dependency subgraph doesn't fail and get removed.

The safest thing is to just always gc() after every ideal build. If it turns out
that this is too costly in performance terms, we can shove it off to a background
process, but it seems to not make much of a difference actually, as long as we
only perform the action once, and not repeatedly.