reify: handle optional fails, create hidden lock
This updates reify to handle any sort of failure in the reification of
any optional node, removing it and the path of optional dependencies
leading up to it, and any nodes made unreachable by the removal.

Also, the resulting actual graph is now stored to
`node_modules/.vlt-lock.json`. It can be loaded by passing
`ignoreHiddenLockfile: false` to the `actual.load()` method. This will
become the default behavior once the loading behavior is updated to
always reload all importers, and only use the hidden lockfile as a
source of information about the contents of `node_modules`.