do lifecycle and chmod at the same time
This is required because we need the dependency bins to be present and
chmodded in order to be able to use them to run a given node's scripts.

So, do the lifecycle scripts and the chmodding in leaf-to-root
topological order (ignoring/breaking cycles, because in that case,
there's just nothing we can logically do to make it work).

This also makes it simpler to eventually detect when a given importer
needs to have its scripts run, because the build's graph-run visit
method can return a flag indicating that the node or one of its
dependencies took some action, and we can also compare this against the
modified times of the contents of the workspace/project, so that
anything that hasn't been touched in some way doesn't have to be
re-built.

For now, though, because we can't possibly know that the workspace even
existed at the last run, we have to just run prepare/install for all
importers no matter what.