change depid delimiter character
The `;;` delimiter is a problem for any child processes that do not
properly escape their paths. This would be something we could maybe get
away with just not supporting, since it's pretty obviously a bug with
the build scripts in question, but unfortunately, `node-gyp` (more
specificlly, `gyp` itself) is one such build tool.

So, use the `·` character as a delimiter instead. It is visually
unassuming (without actually being whitespace), falls within the Latin-1
character set (so not a problem for Windows), and is not a special
character recognized by any shell.

Just in case it ever needs to be adjusted again (and for improved
readability throughout the tests), all instances of literally written
DepIDs have been replaced with `joinDepIDTuple(...)` intead. This also
has the advantage of not requiring us to escape `/` characters as `%2f`.