reify: do not chmod bins to 0o777 (or assume 0o755)
Some systems create files as group-writable by default, and it can be
annoying to not preserve the expected umask. To avoid having to call
deprecated process.umask() method (which is not thread-safe), just
binary-OR 0o111 to whatever mode the file already has, and use that as
the mode for all executables.