build: add a publish script
The publish script relies on the 1Password `op` CLI to get the publish
token and the `npm` CLI to do the actual publishing. In the future This
should change to either a lockfile dependency version of `npm` or the
relevant parts like `libnpmpack`.

This also includes a fix in the build script for finding the correct
relative paths for package.json files in the bundled code, as well as
some refactoring.