semver: Modest performance improvements
This makes a few small improvements to the `new Range()` performance,
and simplifies the `Range.toString()` output.

The creation of a Range object is slightly (~10-20%) slower than the
same operation on `node-semver`, but this is because more work is done
up front in the Range constructor, which `node-semver` leaves until
testing the range against a version (where we are ~40% faster).

Since there is never any reason to create a `Range` object except to
repeatedly test against a parsed `Version`, this is a good tradeoff to
make. The benchmark script is updated to not fail if the Range creation
is slower, as long as Range testing is still faster.