graph: store full spec options to lockfile
Adds spec options (`registry`, `registries`, `scope-registries`,
`git-hosts` and `git-host-archives`) to lockfile, only storing values
that are not part of the default set of values.

Stored values are also used to hydrate a graph that is being loaded from
a lockfile.

Also adds missing fields to json representation of a node (`dev` &
`optional`) along with a missing `projectRoot` property to `GraphLike`
interface.