tar: Do not use worker threads
It seems like the zero-copy ArrayBuffer transfer to node Worker threads
started copying instead at some point, or perhaps is falling over when
it gets to a certain scale. (Possibly resulting from how Buffer slabs
are used internally in Node?)

But! The good news is, it turns out we can go faster by just using
normal async IO and still cut the thread-pool contention down to a
reasonable amount by limiting how many tarballs we unpack in parallel.
So, it's still worth having SOME of the infrastructure that was put in
place for Worker threads, even though it's a wee bit overkill at this
point. But it's fast and very resilient, which is a big plus.