spec: rewrote @vltpkg/spec/browser
`@vltpkg/spec/browser` is now a complete implementation of
`@vltpkg/spec` the missing parts are only the surface that depends on
nodejs APIs, which will only work completely on the non-browser version.

Thinks like unwrapping homedir to resolved path are not available in the
browser / isomorphic version of the package.