registry-client: retry on transient network fails
This fixes an issue where we'd crash when the network is flaky, with
`ECONNREFUSED`, `ECONNRESET`, `EPIPE`, and `UND_ERR_SOCKET`.

Using undici's `RetryHandler` takes care of this transparently,
including retrying `GET` requests with a `Range` header (for servers
that advertise support for it) if the body had already been partly
consumed.