graph: lockfile may load if missing options
Make lockfile load a little more resilient by making sure it does not
throw when trying to deconstruct values from a missing `options`
property.

Even though it should never be the case, we still had some internal
tests that were confused by that, not to mention it might be a user
surface that benefits from the extra resiliency despite not being a file
that should be manually changed.