graph: add flag to track modified deps
Adds a `modifiedDependencies` companion flag to the
`AddImportersDependenciesMap` and `RemoveImportersDependenciesMap`
that tracks if there are any dependencies being add / removed in
any of the importers.

The initial usage of this flag is to skip writing to `package.json`
at the end of a reify in case there were no deps add or removed.

Refs: https://github.com/vltpkg/vltpkg/issues/266