registry-client: remove v8 serialization
Removes the special v8 serialization in favor of using only
standard JSON parse / stringify.

On benchmarks the v8 serialization of json files is not showing
any benefits and this extra info comes at a cost of twice the
cache size for all cached json files and the extra maintainance
complexity.