build: move source rewriting to single esbuild plugin
Previously source rewriting was split between multiple esbuild
plugins which made it easier to read. However I did not realize
that onLoad esbuild plugins will bail once one plugin has claimed
a file. So if multiple plugins were trying to alter the source of
a single file then the first one registered would win.

This caused a recent refactor to no longer copy the gui assets to
the correct location during build because that file already had a
plugin that would operate on it first.

This is fixed by using a single esbuild onLoad plugin that is
responsible for all the source rewriting.