gui(test): quiet console.error for undefined manifests (#365)
I think this only happens in tests but `JSON.stringify(undefined)` is
undefined so when it gets passed to `shiki.codeToHtml` from we get
a lot of console.error output saying the `Cannot read properties of
undefined (reading 'split')`. Also TS doesn't help us here because
JSON.stringify always returns a string according to it. So the easiest
fix is to manually guard this and pass in an empty string instead.