build: run esbuild directly on typescript source
This also cleans up a few build related items:

- Uses esbuild's built-in code splitting. The tradeoff is that we have
to put a few environment variables in the code. It's a bit of a pain but
more explicit and less brittle than the previous approach.
- Uses esbuild's source maps which now map directly back to the TS source.
Previously the sourcemaps were doing an extra hop from TS -> JS -> bundled JS.
This makes stack traces in the bundled published code better and solves the
non-compiled portion of #150.