test: run with node type stripping instead of ts-node
This configures tap to turn off the typescript plugin which uses ts-node
and instead run tests with node's type stripping flags.

When run this way, the tests no longer need tshy to be run during pretest
or presnap events. A future change could remove prepare events also but
for now it is still required for the gui and docs since those look for
the default export condition.