cli: output/view refactoring and test coverage
This is a somewhat significant refactoring of the view/output machinery,
that comes along with a fix for the hang identitfied in #261.

As of this change, cliCommand methods no longer return the result nested
under a `result` key in the returned object. They just return the result
as-is. If they return void, then that's what the return is. To return
either a value or undefined, simply let the return value be something
like `SomeType | undefined`.

This also fixes the strange inconsistency where the `vlt pkg` command
would _always_ output JSON, even in `--view=human` mode. Now, _every_
command will support at minimum `human`, `json`, and `inspect` views. If
`--view=inspect` is set, then the output object will be logged using
`console.log()`.

If a view is specified that is not supported for the given command, then
it will be output using the default view (`human` if stdout is a TTY,
`json` if not).

Similarly, there's no longer a need to nest views under a `{
defaultView, views }` object, or pass around a `isJson` flag. We can
just use the `config.values.view` as the one to use, and if that's
`json`, then that's how we know it's JSON.

This is _not_ quite the permanent solution, because the ideal way to fix
the hang is to pass an AbortSignal object through the tree of requests,
so that it does not forcibly exit the process if for example an install
is performed from the GUI. The AbortController/AbortSignal fix will come
in the next commit.

Fix: #261