read-project-folders: improve folder ignoring and descent
This is a performance improvement, and cuts out some common sources of an
explosion of unnecessary folder descent in reading the project folders.

Note: the intent is to eventually replace this ignore-list with some UX to
do opt-in for folder traversal, but that needs to be built, and will be much
easier once the gui-server functionality is moved out of cli.