dev: remove prepare scripts and point default exports at .ts files (#446)
`tshy`-able workspaces no longer need preparing. This is the result of all the
previous PRs running TypeScript source directly. The `docs` and `gui` still have
prepare scripts since those still need preparing for the web.

The `docs` prepare script has been sped up so it doesn't run `typedoc` anymore.
For docs it only needs to build the astro typedefs so that editors/linting have
those available. The full build process for the docs now only runs during the
build on Vercel.

---

This also includes a patched version of tshy to make the default export
condition point directly at the TypeScript source and not actually build
anything to ./dist. This allows us to remove all the custom config tooling
conditions and needing to set --condition in NODE_OPTIONS. This patch should get
removed if this or a similar feature lands in tshy:
https://github.com/isaacs/tshy/issues/101

tshy will still do the right thing and build/export ./dist files when run during
workspace pack or publish. We're not actively publishing our workspaces but it's
nice that this will Just Work when we do.