package-info, registry-client: integrity checking
This updates the RegistryClient and PackageInfoClient to check integrity
in cases where it's not already coming from a trusted source, and to
raise an error when an expected integrity does not match what we
receive.

The only case where we actually have to *calculate* an integrity value
is when we have a lockfile, but a cold cache and empty `node_modules`.
In that case, it's around 100x faster to just calcualte the integrity
than to fetch it from the registry, since we otherwise do not need to
fetch the manifest, as we have a lockfile already.

This also sends the `accept-integrity` request header, and handles the
`integrity` response header, so that integrity checking can be performed
on the server instead of the client, which would be more efficient,
though it is not supported by the public npm registry.