dev: remove unused code and dependencies with knip (#546)
This PR adds and configures [`knip`](https://knip.dev) for the monorepo.
Knip will statically analyze a project and list files, dependencies,
exports, types, etc that are unused. This PR only focuses and files and
dependencies because I manually found some unused of both.

This is not adding a CI check and I plan to periodically run this
locally.