registry-client: stale-while-revalidate behavior
If a request is older than the cache-control maxage, but younger than
either the `stale-while-revalidate` cache-control value if it exists, or
60 times the cache-control maxage (configurable via the
`stale-while-revalidate-factor` option), then we will return the stale
cached entry and revalidate it in the background.

This addresses the necessarily quite short cache TTLs on packuments,
while still providing optimal performance in most cases that a user is
likely to encounter.