gui: dont mark as private in package.json
Previously the GUI was marked as private so that it wouldn't appear in the
generated typedoc docs. But now that we are publishing workspaces, if the
GUI is private then anything that depends on it would need to be private
also. It's easier to make everything public and publishable and not worry
about it.

The GUI is now excluded from generated docs since it has no observable
exports.