allow calling spawned procs from compiled bins (#553)
This changeset creates a new pattern in the bins to allow calling other
entry points (such as cache-unzip) when compiled.

For now this pattern is only applied to cache-unzip but will be used for
all other spawned processes in the future.

Some notes:
- Does not work on Windows because of the lack of detached process
support on the platform in Deno
(https://github.com/denoland/deno/issues/25867). On Windows the main
process will remain open while the cache is being unzipped.
- There is a smoke test to ensure the cache is always unzipped, but
there is a failure on Windows that is being skipped for now. A follow-up
PR will be created to debug and fix that test.