cache-unzip: write unzipped result to existing integrity files (#622)
With this change the `cache-unzip` process will now link any existing
integrity files to the unzipped cache entry.

I tried a different approach to this originally where
`@vltpkg/cache.set` was responsible for updating the integrity files.
But we already use `cache.set(key, val, { integrity })` to mean "write
this value to the cache, first linking *from* the integrity file if it
exists" which is the behavior we want everywhere else.

So rather than make a new method/option on `@vltpkg/cache` for this less
used behavior, `cache-unzip` uses the `cache.path()` and
`cache.integrityPath()` methods to manually re-link any integrity files.

Fixes #617