query: add score selector
Adds a `:score("<rate>", [kind])` pseudo selector that matches packages
based on the scores rated found in the insights data.

The `rate` parameter can be prefixed with a comparator (>, < , >=, <=,
similar to the `:published` selector in order to compare against a
range instead of comparing exact values. If no comparator is provided,
it will match exact scores.

Note that `rate` values are percentual values and can be either 0-1 or
0-100 and will be normalized prior to comparison.

Examples:

    /* Matches packages with exactly 0.8 overall score */
    :score(80)

    /* Matches packages with overall score greater than 0.8 */
    :score(">0.8")

    /* Matches maintenance score less than or equal to 0.5 */
    :score("<=0.5", "maintenance")

Fixes: https://github.com/vltpkg/statusboard/issues/108