cli-sdk: more comprehensive error messaging (#628)
#579 without any of the significant changes to `error-cause`.

If we are printing any `Error` without a handled code (whether from
`error-cause` or not), we now print all the information from the error
including the stack. But we cap each printed bit of information at 200
lines to avoid blowing up the terminal.

Also uses `util.format` everywhere with common options from `view` so we
get colors.