query: fix unmatching pseudo selectors (#689)
When in loose mode, unknown pseudo selectors names will now fail to
match any item. Using multiple selectors in the root query, e.g: `#a,
#b, #c` will now switch loose mode on by default to the remainder of the
query parser.

This makes it so that if the user has an unidentified pseudo selector in
a multiple-selector query, e.g: `#a, :foo` they will still get a result
with items matching `#a` while `:foo` is ignored.