query: refactor project selector (#694)
Tweaks the `:project` pseudo selector so that it now behaves similar to
other selectors as a filter instead of always returning all importer
nodes no matter what.

This will fix problems in which the selector would not work properly if
not used as the first element,e.g:

    vlt query '#a:project'

    vs

    vlt query ':project#a'

With this refactor both examples above will always return the same
values.