cli-sdk: use util.styleText over chalk (#692)
The reason for this change is not necessarily replacing `chalk` with
`util.styleText`, but instead to decouple checking color support with
the writing of ansi color sequences.

With this change we now detect whether both `stderr` and `stdout` should
be colorized. And `ViewOptions` now only contains a `colors` boolean
indicating support for `stdout` since we assume that our views will be
writing to `stdout` primarily.

As a result, producers of colorized output (eg `graph`) can now use any
color library without worrying about detecting support since that is all
done in `output.ts`.

There is one `TODO` left which is that `output` exports mutable
`styleText` functions. Those are only used by `exec` and will be cleaned
up as part of #684.